Add type stubs
Add type stubs for the library as a separate stub file which means we can
use the python 3 syntax, with the downside of having to keep them in
sync with the source.

I ran [monkeytype][1] and [stubgen][2] over responses.py and combined the
results manually.

I had to move the module into a package for the types to work.
see: https://www.python.org/dev/peps/pep-0561/#packaging-type-information

Fixes https://github.com/getsentry/responses/issues/339

https://www.python.org/dev/peps/pep-0561/

[1]: https://github.com/Instagram/MonkeyType
[2]: https://mypy.readthedocs.io/en/stable/stubgen.html