Add flag to allow automatic calculation of the Content-Length header (#398)
To be correct, the Content-Length header must be calculated from the
binary representation of the body. This is calculated as late as
possible by calling _handle_body(), and this is all done privately
within the module. Therefore, it's easiest if this is done in a
centralised place as late as possible by this module.

The new behaviour is controlled by a flag to avoid breaking backwards
compatibility. There are still plenty of situations where you wouldn't
want the Content-Length header to be set.