recorder: use a case-insensitive HTTP headers check
_recorder._remove_default_headers() does a case-sensitive check for
HTTP header field names are case insensitive. The result is that for
example, a lower-case 'content-type' field will not be removed. Later,
when _add_from_file tries to load the data using RequestsMock.add(), a
RuntimeError exception will be raised when it finds the 'content_type'
kwarg was passed and 'content-type' is in the 'headers' kwargs.

Fix this by simply having _remove_default_headers() do a
case-insensitive check.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>